Project renamed and scripts subpackage removed.
Project name changed to Aracne to avoid conflicts with the locate UNIX utility.
The new name will be used as codename of the project or maybe it will stay as
the final name.

The scripts subpackage was removed becouse it's functionality will be
implemented in the script installed by the setup.py.