Created skeleton of execution.
Created new Controller component (in a controller.py aracne submodule) wich is
responsible for creating, starting and stopping the rest of the components.  It
subclass the Daemon class.

Implemented the run() method of the CrawlManager and ProcessorManager that both
runs on an independent thread of execution processing items until the running
flag is cleared.

Created initializer of the TaskQueue and ResultQueue classes required to test
the execution of the other components (specially the Controller).

Changes made to the aracne script to create the Controller.