Implemented processor main loop and other changes.
Created aracne.errors module to hold all the exceptions of the package.

Added process() and __init__() methods to the ResultProcessor abstract class
and created a NaiveProcessor that will be the default processor of the crawler
(its process() method that currently does nothing).

Implemented the ResultQueue using a Queue.Queue instance internally.  This
needs to be changed to save the crawl results waiting to be processed using
other way (maybe in HDD).

Added the 'processor' option (required by the ProcessorManager) to the
'processormanager' configurations.