Multiple changes.
Changed the aracne script to support the structue that will have the
configuration file of the application.  Removed --pidfile option, this will be
specified in the configuration file.  Added DEFAULT_LOGFILE that will be used
by the _parse_configfile method when it supports reading the configuration
file.  Modified _parse_configfile to return a dictionary that matches the new
structure.

Updated Controller, CrawlerManager and ProcessorManager that used old keys from
the config dictionary to match the new structure.

Internal variables renamed in Controller to have short names. Docstrings change
and other minor changes.