Implemented CrawlerManager and (partially) SiteCrawler.
Implemented the CrawlerManager and the main loop of the SiteCrawler (executing
the crawl tasks is still not implemented).  The TaskQueue was modified to
support the new structure of the SiteCrawler which will be responsible of
sleeping the thread if no task is available.

Renamed _crawler attribute to _crawlers in the Controller because the
CrawlerManager usually will manage more that one SiteCrawler.