Site ids initialized, set version number and other changes.
Modified AracneDaemon to initialize the ids and create instances of URL for the
root URL of each site.

Version number of the project set to 0.1.0.  When it gets ready it will be the
first public release.

Added a log message showing the number of configured sites and other log
messages were modified.