Source code refractoring.
Multiple changes.  Changed order of the arguments of the initilizers of various
classes to have the same other that other related components.  Modified daemon
subclasses to receive the arguments instead of a config dictionary with the
values.  Modified names of options in the configuration files.  Docstrings
modified and comments added.