Bugs fixed.
Moved initialization of the TaskQueue and ResultQueue to the run() method of
the ArachneDaemon class because Berkeley DB handlers should not be shared
across process forks.

Added DB_THREAD flag to the Berkeley DB environment and database handlers
because they will be used in the context of more than one thread.

Added missing / in setup.py.

Fixed wrong spelled self._results FTPHandler member.