TaskQueue initialization now checks for site id in the database.
When the application is in the middle of the execution of a task the site id is
out of the sites.db database.  Therefore, if the application crashes in this
moment the site is lost.  Now in the __init__ method checks that the site ids
of the configured sites are in site.db.