XapianProcessor modified.
dirname will be stored as a value becouse it could be longer than the maximum
length of the terms.  XapianProcessor modified to use queries with the
OP_VALUE_RANGE operator to match the sub-directories.

Changed term prefixes, basename modified to be a a value instead of a term,
added value for the is_dir metadata and other minor changes.