Minor modifications to IndexProcessor.
Modified the __init__ method to receive the path to the database directory
instead of the directory where the index should be created. This is required
to create IndexSearcher in the future. ArachneDaemon class updated.

Changed visibility of some constants defined as attributes of the
IndexProcessor class (mainly term prefixes and value slots).