Created initial version of the search form and a bug fixed.
A first version of the search form was created (basic and advanced) added the
CSS file, the big logo image and the required templates. The unused templated
were removed.

Fixed bug in the IndexProcessor that caused the entry for the root directory in
the index to be removed when the result for the root directory itself was
processed.